New OpenFile test focusing on the stack overflow question 66885882
Writing a streamed cross-reference in PDF: file is detected as damaged?
https://stackoverflow.com/questions/66885882/writing-a-streamed-cross-reference-in-pdf-file-is-detected-as-damaged
corrupted-pdf-2.pdf
https://easyupload.io/p5352r

testCorruptedPdf2 - Fixing the issues found in testCorruptedPdf() and testCorruptedPdfStartxref() resulted in this new test file which still fails to be read: The OP applied prediction only to the offset bytes, not the type byte. Thus, after resolving prediction the types of the entries were mostly wrong.