New AnnotateSignedPdf test focusing on the stack overflow question 71005453
"At least one signature is invalid" After adding stamp in Signed PDF
https://stackoverflow.com/questions/71005453/at-least-one-signature-is-invalid-after-adding-stamp-in-signed-pdf

testAddImageStamp - This test illustrates how to add a stamp annotation with an image to a signed PDF without invalidating the signature. In contrast to the OP's code we don't use the OverContent of a page to create a PdfAppearance instance but instead use the static method PdfAppearance.createAppearance. Thus, we don't trigger the side effects of the OverContent retrieval.