New StampTextOnTagged test focusing on the stack overflow question 76388538
Cannot write a paragraph to a pdf file using iText pdf
https://stackoverflow.com/questions/76388538/cannot-write-a-paragraph-to-a-pdf-file-using-itext-pdf
test_file.pdf
https://drive.google.com/file/d/1d2PKbf1pqXwZTboLZOJr69KJTJtAKYUx/view

testLikeFrançoisVinckel:

This is essentially the OP's code. Running it with iText 5.5.4 one indeed runs into a NullPointerException. Running it with iText 5.5.13.3, though, the code executes without issue.

Comparing the exception location code with its current form one finds that there indeed was an unconditional call of a method of a parent object but that now there is a call to a helper method that first checks the parent and only calls its method if it isn't null.

The fix was committed by Alexander Chingarev on 2015-01-19 14:39:23 with the message "Fixed NPE when modifying content of TaggedPDF document." (e0eec31b31a414ad4fe70eda1aeea04fc25dd319)