New AddWatermark test focusing on the stack overflow question 78135326
Unable to make itext5 pdf watermark non removable in VMware Workspace ONE Boxer email
https://stackoverflow.com/questions/78135326/unable-to-make-itext5-pdf-watermark-non-removable-in-vmware-workspace-one-boxer

testWatermark1LikeMahmoudSaleh - In contrast to what the OP said, Acrobat can quite easily remove this watermark, simply using the "Edit a PDF" tool, one selects the page-covering object and deletes it. This has to be done some 35 times because the code fills the overcontent that many times with the pattern, but thereafter the mark is gone.