New AddWatermark test focusing on the stack overflow question 78135326
Unable to make itext5 pdf watermark non removable in VMware Workspace ONE Boxer email
https://stackoverflow.com/questions/78135326/unable-to-make-itext5-pdf-watermark-non-removable-in-vmware-workspace-one-boxer

testWatermarkAllInPattern - As proposed in a comment, one can try to put all the former page content into the pattern, so that Acrobat removes the content together with the watermark. It turned out, though, that Acrobat suddenly allows editing the pattern content if there is no other content in the page than the pattern. Thus, here we also add a short pseudo content.

Beware, chances are that Acrobat will eventually also allow to edit this kind of watermarking. Maybe it even now is possible, merely not as obvious as before.